+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # Unreleased
 
 ### Added
-- Support renderOption prop in Autocomplete #604 by @ihor-lazariev
+- Support `renderOption`, `clearButtonProps` and `clearable`  props in Autocomplete #604 by @ihor-lazariev
 
 # 2.0.0